Add replace support to SQLite adapter
